+++ include/dcc/rx/east_west.hpp
@@ -16,7 +16,7 @@
 namespace dcc::rx {
 
 template<typename T>
-concept EastWest = requires(T t, uint32_t addr, std::optional<bool> opt_dir) {
+concept EastWest = requires(T t, uint16_t addr, std::optional<bool> opt_dir) {
   { t.eastWestDirection(addr, opt_dir) } -> std::same_as<void>;
 };
